+++ nope/lib/models/Media.php
@@ -5,13 +5,18 @@ namespace Nope;
 use RedBeanPHP\R as R;
 use Respect\Validation\Validator as v;
 use Respect\Validation\Exceptions\NestedValidationException;
+use Intervention\Image\ImageManagerStatic as Image;
 
 class Media extends Content {
 
   const MODELTYPE = 'media';
 
+  function isExternal() {
+    return !is_null($this->provider);
+  }
+
   function getPath() {
-    return NOPE_UPLOADS_DIR . $this->model->filename;
+    return NOPE_UPLOADS_DIR . $this->filename;
   }
 
   function getUrl($cache=true) {
@@ -19,10 +24,31 @@ class Media extends Content {
     return NOPE_UPLOADS_PATH . $this->filename . $t;
   }
 
+  function getAbsoluteUrl($cache=true) {
+    if($this->isExternal()) {
+      return $this->url;
+    }
+    return Utils::getBaseUrl() . $this->getUrl($cache);
+  }
+
   function isImage() {
+    if($this->isExternal()) {
+      return $this->type === 'image';
+    }
     $needle = 'image/';
     $length = strlen($needle);
     return (substr($this->mimetype, 0, $length) === $needle);
+  }
+
+  function getMetadata() {
+    if($this->isImage()) {
+      $path = $this->getPath();
+      return (object) [
+        'exif' => Image::make($path)->exif(),
+        'iptc' => Image::make($path)->iptc()
+      ];
+    }
+    return null;
   }
 
   function delete() {
@@ -50,20 +76,30 @@ class Media extends Content {
     }
   }
 
+  function beforeSave() {
+    $this->metadata = json_encode($this->getMetadata());
+    parent::beforeSave();
+  }
+
   function jsonSerialize() {
-    $obj = parent::jsonSerialize();
-    $obj->url = $this->getUrl();
-    $obj->preview = (object) [];
+    $json = parent::jsonSerialize();
+    $json->url = $this->getUrl();
+    $json->absoluteUrl = $this->getAbsoluteUrl();
+    $json->preview = (object) [];
     foreach (\Nope::getConfig('nope.media.size') as $key => $value) {
-      $obj->preview->$key = $this->getPreview($key);
+      $json->preview->$key = $this->getPreview($key);
     }
-    $obj->isImage = $this->isImage();
-    unset($obj->cover);
-    return $obj;
+    $json->isImage = $this->isImage();
+    $json->isExternal = $this->isExternal();
+    $json->metadata = json_decode($this->metadata);
+    $json->starred = (bool) $json->starred;
+    unset($json->cover);
+    return $json;
   }
 
   function validate() {
-    $contentValidator = v::attribute('title', v::length(1,255));
+    $contentValidator = v::attribute('title', v::length(1,255))
+      ->attribute('filename', v::length(1,255));
     try {
       $contentValidator->check((object) $this->model->export());
     } catch(NestedValidationException $exception) {
@@ -72,24 +108,54 @@ class Media extends Content {
     return true;
   }
 
-  static public function findById($id) {
-    return self::__to(R::findOne(self::MODELTYPE, 'id = ?', [$id]));
-  }
-
-  static public function findAll($filters=null, $limit=-1, $offset=0, &$count=0, $orderBy='id desc') {
-    $filters = (object) $filters;
+  static public function findAll($filters=[], $limit=-1, $offset=0, &$count=0, $orderBy='starred desc,id desc') {
     $params = [];
-    /*if($filters->role) {
-      $sql[] = 'role = ?';
-      $params[] = $filters->role;
-    }*/
+    $sql = self::__getSql($filters, $params);
     if($orderBy) {
       $sql[] = 'order by '.$orderBy;
     }
-    $users = R::findAll(self::MODELTYPE, implode(' ',$sql),$params);
-    return self::__to($users, $limit, $offset, $count);
+    $contentsList = R::findAll(self::__getModelType(), implode(' ',$sql),$params);
+    return self::__to($contentsList, $limit, $offset, $count);
   }
 
-
+  static function __getSql($filters, &$params=[], $p = null) {
+    $sql = [];
+    $filters = (object) $filters;
+    if($filters->author) {
+      $sql[] = $p.'author_id = ?';
+      $params[] = $filters->author->id;
+    }
+    if($filters->excluded) {
+      if(count($sql)) {
+        $sql[] = 'and';
+      }
+      $sql[] = $p.'id NOT in (' . R::genSlots($filters->excluded) . ')';
+      foreach ($filters->excluded as $value) {
+        $params[] = $value;
+      }
+    }
+    if($filters->mimetype) {
+      if(count($sql)) {
+        $sql[] = 'and';
+      }
+      if(substr($filters->mimetype,0,1)==='!') {
+        $sql[] = $p.'mimetype NOT LIKE ?';
+        $filters->mimetype = substr($filters->mimetype,1);
+      } else {
+        $sql[] = $p.'mimetype LIKE ? and provider is null';
+      }
+      $params[] = '%' . $filters->mimetype . '%';
+    }
+    if($filters->text) {
+      if(count($sql)) {
+        $sql[] = 'and';
+      }
+      $like = '%' . $filters->text . '%';
+      $sql[] = '(title LIKE ? or description LIKE ?)';
+      $params[] = $like;
+      $params[] = $like;
+    }
+    return $sql;
+  }
 
 }