+++ nope/lib/views/page/form.php
@@ -6,7 +6,7 @@
           <div class="panel-body">
             <div class="toolbar">
               <a ng-click="content.starred=!content.starred;" class="btn btn-star pull-right"><i class="fa" ng-class="{'fa-star-o':!content.starred,'fa-star':content.starred}"></i></a>
-              </div>
+            </div>
             <div class="form-group" ng-class="{'has-error':(!contentForm.title.$valid && contentForm.title.$touched)}">
               <label class="control-label">Title</label>
               <input type="text" name="title" class="form-control input-lg" ng-model="content.title" required />
@@ -16,7 +16,7 @@
               <textarea name="body" class="form-control" ng-model="content.body" rows="20"></textarea>
             </div>
             <div class="form-group">
-              <label>Tags</label>
+              <label>Tags (comma separated)</label>
               <input type="text" name="tags" class="form-control" ng-model="content.tags" />
             </div>
           </div>
@@ -26,8 +26,8 @@
         <div class="panel panel-default">
           <div class="panel-heading content-author">
             <div class="list-group">
-              <div class="list-group-item clearfix">
-                <img ng-src="{{content.author.cover.preview.icon}}" class="img-circle" />
+              <div class="list-group-item clearfix" ng-class="{'has-image':content.author.cover}">
+                <img ng-src="{{content.author.cover.preview.icon}}" class="img-circle" ng-if="content.author.cover" />
                 Created by <span class="fullname">{{content.author.prettyName || content.author.username}}</span>
                 <span ng-if="content.id">{{content.creationDate | nopeMoment:'fromNow'}}</span>
                 <span ng-if="!content.id">now</span>
@@ -39,7 +39,7 @@
               <div class="preview-image" ng-show="content.cover" ng-style="{backgroundImage:'url('+content.cover.preview.thumb+')'}">
                 <a href="" ng-click="content.cover=null" class="btn btn-danger btn-xs"><i class="fa fa-times-circle"></i></a>
               </div>
-              <nope-model model="Media" ng-model="content.cover" multiple="false" preview="icon" label="Add featured image"></nope-model>
+              <nope-model href="#/media?mimetype=image/" ng-model="content.cover" multiple="false" label="Add featured image"></nope-model>
             </div>
             <div class="form-group" ng-class="{'has-error':(!contentForm.slug.$valid && contentForm.slug.$touched)}">
               <label class="control-label">Slug</label>