+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## v0.5.0 (2016-01-18)
+
+### Libraries
+- Added **Embed** version 2.6.1.
+
+### Breaking changes
+- `Content::setTags` is now a public method.
+
+### Features
+- Better and more complete `Media` model with corresponding UI.
+- Import media from external providers.
+- Open contents list inside modal for selection.
+
+### Fixes
+- Fixed wrong `Page` edit link on content detail.
+- Fixed default `priority` on `Page` creation.
+- Fixed missing `starred` field on `media` table during development phase.
+
 ## v0.4.0 (2016-01-13)
 
 ### Libraries